+++ NWN.Anvil/src/main/API/Variable/ObjectStorage/ObjectStorageVariable.cs
@@ -1,0 +1,13 @@
+namespace Anvil.API
+{
+  public abstract class ObjectStorageVariable<T> : ObjectVariable<T>
+  {
+    protected string Key => VariableTypePrefix + Name;
+
+    protected abstract string VariableTypePrefix { get; }
+
+    protected abstract string ObjectStoragePrefix { get; }
+
+    protected abstract bool Persist { get; }
+  }
+}