+++ NWN.Anvil/src/main/API/Variable/ObjectStorage/ObjectStorageVariableFloat.cs
@@ -1,0 +1,43 @@
+using Anvil.Services;
+
+namespace Anvil.API
+{
+  public abstract class ObjectStorageVariableFloat : ObjectStorageVariable<float>
+  {
+    [Inject]
+    private static ObjectStorageService ObjectStorageService { get; set; }
+
+    public sealed override bool HasValue => ObjectStorageService.TryGetObjectStorage(Object, out ObjectStorage objectStorage) && objectStorage.ContainsFloat(ObjectStoragePrefix, Key);
+
+    public sealed override float Value
+    {
+      get => ObjectStorageService.GetObjectStorage(Object).GetFloat(ObjectStoragePrefix, Key).GetValueOrDefault();
+      set => ObjectStorageService.GetObjectStorage(Object).Set(ObjectStoragePrefix, Key, value, Persist);
+    }
+
+    protected sealed override string VariableTypePrefix => "PERFLT!";
+
+    public sealed override void Delete()
+    {
+      ObjectStorageService.GetObjectStorage(Object).Remove(ObjectStoragePrefix, Key);
+    }
+  }
+
+  public sealed class PersistentVariableFloat : ObjectStorageVariableFloat
+  {
+    protected override bool Persist => true;
+    protected override string ObjectStoragePrefix => "NWNX_Object";
+  }
+
+  internal sealed class InternalVariableFloat : ObjectStorageVariableFloat
+  {
+    protected override bool Persist => true;
+    protected override string ObjectStoragePrefix => "ANVIL_API";
+
+    internal sealed class Persistent : ObjectStorageVariableFloat
+    {
+      protected override bool Persist => false;
+      protected override string ObjectStoragePrefix => "ANVIL_API";
+    }
+  }
+}